refactor(dock): extract item model
