+++ meson.build
@@ -459,6 +459,7 @@ _noctalia_sources = files(
   'src/shell/dock/dock_context_menu.cpp',
   'src/shell/dock/dock_geometry.cpp',
   'src/shell/dock/dock_instance.cpp',
+  'src/shell/dock/dock_items.cpp',
   'src/shell/desktop/desktop_widget.cpp',
   'src/shell/desktop/desktop_widget_factory.cpp',
   'src/shell/desktop/desktop_widget_settings_registry.cpp',
